refactor: split HomePage into components and move to components folder
