+++ src/pages/home/components/Brands.tsx
@@ -1,0 +1,5 @@
+import styles from './../HomePage.module.css';
+
+export function Brands() {
+  return <h2 className={styles.brands}>Brands</h2>;
+}
