+++ src/pages/home/components/Recommend.tsx
@@ -1,0 +1,5 @@
+import styles from './../HomePage.module.css';
+
+export function Recommend() {
+  return <h2 className={styles.recommend}>Our Recommendations</h2>;
+}
